preliminary work
