+++ web/src/engine/websites/InariManga.ts
@@ -1,10 +1,11 @@
 import { Tags } from '../Tags';
-import icon from './LectorJPG.webp';
+import icon from './InariManga.webp';
 import { Kozaku } from './templates/Kozaku';
+
 export default class extends Kozaku {
 
     public constructor() {
-        super('lectorjpg', 'LectorJPG', 'https://lectorjpg.com', Tags.Media.Manhwa, Tags.Media.Manga, Tags.Language.Spanish, Tags.Source.Aggregator);
+        super('inarimanga', 'InariManga', 'https://inarimanga.rzhost.xyz', Tags.Media.Manhwa, Tags.Media.Manga, Tags.Language.Spanish, Tags.Source.Aggregator);
     }
 
     public override get Icon() {
