+++ web/src/engine/websites/_index.ts
@@ -57,6 +57,7 @@ export { default as Boomtoon } from './Boomtoon';
 export { default as CartelDeManhwas } from './CartelDeManhwas';
 export { default as CarToonMad } from './CarToonMad';
 export { default as CasaComic } from './CasaComic';
+export { default as CatharsisWorld } from './CatharsisWorld';
 export { default as CatTranslator } from './CatTranslator';
 export { default as Chochox } from './Chochox';
 export { default as CiaoPlus } from './CiaoPlus';
@@ -224,6 +225,7 @@ export { default as IllusionScan } from './IllusionScan';
 export { default as Imgur } from './Imgur';
 export { default as IMHentai } from './IMHentai';
 export { default as ImperioDaBritannia } from './ImperioDaBritannia';
+export { default as InariManga } from './InariManga';
 export { default as InfraFandub } from './InfraFandub';
 export { default as InManga } from './InManga';
 export { default as InmortalScan } from './InmortalScan';
